refactor: extract `SignupDisclaimer` to separate component file
