+++ packages/shared/src/components/auth/EmailSignupForm.tsx
@@ -1,48 +1,11 @@
 import React, { ReactElement, useState } from 'react';
-import classNames from 'classnames';
 import { Button, ButtonSize } from '../buttons/Button';
 import { TextField } from '../fields/TextField';
 import ArrowIcon from '../icons/Arrow';
 import MailIcon from '../icons/Mail';
 import AuthForm from './AuthForm';
-import { privacyPolicy, termsOfService } from '../../lib/constants';
 import { IconSize } from '../Icon';
-
-interface SignupDisclaimerProps {
-  className?: string;
-}
-export function SignupDisclaimer({
-  className = '',
-}: SignupDisclaimerProps): ReactElement {
-  return (
-    <p
-      className={classNames(
-        'w-full text-center text-theme-label-quaternary typo-caption1',
-        className,
-      )}
-    >
-      By signing up I accept the{' '}
-      <a
-        href={termsOfService}
-        target="_blank"
-        rel="noopener"
-        className="font-bold underline"
-      >
-        Terms of Service
-      </a>{' '}
-      and the{' '}
-      <a
-        href={privacyPolicy}
-        target="_blank"
-        rel="noopener"
-        className="font-bold underline"
-      >
-        Privacy Policy
-      </a>
-      .
-    </p>
-  );
-}
+import SignupDisclaimer from './SignupDisclaimer';
 
 interface EmailSignupFormProps {
   onSubmit: (e: React.FormEvent) => unknown;
